Use unitless `line-height`; fix footnotes (v0.1.3)
Specifying `line-height: 1.3em` instead of simply `line-height: 1.3`
causes odd inheritance properties. In the case of the former, the
`line-height` is calculated from the `font-size` of the parent only. In
the case of the later, the `line-height` is calculated from the
`font-size` of the current child element.

This was causing odd spacing issues with footnotes, as the computed
`line-height` was far too large, since footnotes use a really small
`font-size`.

A new test is included for footnotes and additional tests are included
in the inline code-block test in order to test multiple font sizes.